[REFACT] 멤버 서비스 인터페이스 삭제
